backend/python/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-dotenv
pydantic
requests
aiohttp
typing-extensions
python-jose
passlib
bcrypt
pandas
numpy
scikit-learn
mysqlclient
python-multipart
pytest
black
isort
flake8 